test(byoo): move performance suite under tests
