+++ src/compute-plane-services/byoo-otel-collector/Makefile
@@ -1,4 +1,4 @@
-.PHONY: test perf-test docker-build docker-run update-examples update-config-template validate-config
+.PHONY: test docker-build docker-run update-examples update-config-template validate-config
 
 DOCKER_TAG := $(shell whoami)-dev
 OTEL_BUILDER_VERSION ?= v0.157.0
@@ -11,13 +11,6 @@ OTEL_BUILDER_VERSION ?= v0.157.0
 # does not contain modules listed in go.work" failure from the perf dir.
 test:
 	find . -name go.mod -execdir env GOWORK=off go test -v ./... \;
-
-# Build, vet, and test the standalone perf module. The perf module is not part
-# of go.work and has no Bazel targets, so it is invisible to the Bazel CI lane;
-# this target is wired into the required byoo-otel-collector CI workflow so the
-# perf tests actually run on every change to this subtree.
-perf-test:
-	cd perf && GOWORK=off go build ./... && GOWORK=off go vet ./... && GOWORK=off go test ./...
 
 update-examples:
 	./scripts/update-examples.sh
@@ -33,4 +26,3 @@ docker-build:
 
 docker-run:
 	docker run --network=host -v./accounts-secrets.json:/var/secrets/accounts-secrets.json -v./secrets:/etc/byoo-otel-collector/secrets/ -v./test/local/otelconfig.yaml:/etc/byoo-otel-collector/config.yaml byoo-otel-collector:$(DOCKER_TAG) $(EXTRA_ARGS)
-
