+++ tests/perf/byoo-otel-collector/pkg/k3d/k3d_test.go
