+++ tests/perf/byoo-otel-collector/pkg/labels/labels.go
