+++ tests/perf/byoo-otel-collector/pkg/profile/profile.go
