+++ tests/perf/byoo-otel-collector/pkg/profile/profile_test.go
