+++ tests/perf/byoo-otel-collector/pkg/render/render_test.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/NVIDIA/nvcf/src/libraries/go/lib/pkg/icms-translate/translate/common"
 
-	"github.com/NVIDIA/nvcf/src/compute-plane-services/byoo-otel-collector/perf/pkg/spec"
+	"github.com/NVIDIA/nvcf/tests/perf/byoo-otel-collector/pkg/spec"
 )
 
 func TestRenderContainerExtractsSidecar(t *testing.T) {
