+++ tests/perf/byoo-otel-collector/pkg/sink/sink_test.go
@@ -23,7 +23,7 @@ import (
 
 	"sigs.k8s.io/yaml"
 
-	"github.com/NVIDIA/nvcf/src/compute-plane-services/byoo-otel-collector/perf/pkg/labels"
+	"github.com/NVIDIA/nvcf/tests/perf/byoo-otel-collector/pkg/labels"
 )
 
 const testNS = "byoo-perf"
