+++ tests/perf/byoo-otel-collector/pkg/spec/spec.go
