+++ tests/perf/byoo-otel-collector/pkg/spec/spec_test.go
